+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Unreleased
 
 * Upgrade the `org.realityforge.keycloak.client.authfilter:keycloak-jaxrs-client-authfilter` artifact to version `1.01`.
+* Add local validation to ensure top-level fields inside operations are defined by the schema.
 
 ### [v0.06](https://github.com/realityforge/giggle/tree/v0.06) (2019-08-19)
 [Full Changelog](https://github.com/realityforge/giggle/compare/v0.05...v0.06)